Move modules tools.py to analysis.py
