+++ bin/make_primary_goes_list
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-import fetchsep.utils.tools as tools
+import fetchsep.utils.analysis as tools
 import argparse
 
 
@@ -37,11 +37,33 @@ parser.add_argument("--Files", type=str, default='',
 parser.add_argument("--Prefix", type=str, default='GOES',
         help=("The final SEP list will be named like prefix_PRIMARY.YYYY-MM-DD.YYYY-MM-DD_sep_events.csv. Default=GOES"))
 
+parser.add_argument("--datapath", type=str, default=None,
+        help=("Specify path where satellite data should be downloaded. Will default to "
+            "datapath listed in fetchsep.cfg if a value is not specified."))
+
+parser.add_argument("--outpath", type=str, default=None,
+        help=("Specify path where output files should be saved. Will default to "
+            "outpath listed in fetchsep.cfg if a value is not specified."))
+            
+parser.add_argument("--plotpath", type=str, default=None,
+        help=("Specify path where plots should be saved. Will default to "
+            "plotpath listed in fetchsep.cfg if a value is not specified."))
+
+parser.add_argument("--listpath", type=str, default=None,
+        help=("Specify path where lists should be saved. Will default to "
+            "listpath listed in fetchsep.cfg if a value is not specified."))
+
+
 args = parser.parse_args()
 
 filename = args.Filename
 files = args.Files
 prefix = args.Prefix
+
+path_to_data = args.datapath
+path_to_output = args.outpath
+path_to_plots = args.plotpath
+path_to_lists = args.listpath
 
 lists = ''
 if filename != '':
@@ -50,6 +72,10 @@ elif files != '':
     files = files.strip().split(",")
     lists = files
 
-tools.create_primary_goes_sep_list(lists, prefix=prefix)
+tools.create_primary_goes_sep_list(lists, prefix=prefix,
+    path_to_data=path_to_data,
+    path_to_output=path_to_output,
+    path_to_plots=path_to_plots,
+    path_to_lists=path_to_lists)
